fix: background pid refresh
