+++ apps/easypid/src/app/(app)/_layout.tsx
@@ -10,7 +10,6 @@ import { type CredentialDataHandlerOptions, DeeplinkHandler, useHaptics } from '
 import { HeroIcons, IconContainer } from '@package/ui'
 import { useEffect, useState } from 'react'
 import { useTheme } from 'tamagui'
-import { WithBackgroundPidRefresh } from '../../features/pid/WithBackPidRefresh'
 
 const jsonRecordIds = [activityStorage.recordId]
 
@@ -99,48 +98,46 @@ export default function AppLayout() {
   return (
     <AgentProvider agent={secureUnlock.context.agent}>
       <WalletJsonStoreProvider agent={secureUnlock.context.agent} recordIds={jsonRecordIds}>
-        <WithBackgroundPidRefresh>
-          <Stack screenOptions={{ headerShown: false }}>
-            <Stack.Screen
-              options={{
-                presentation: 'modal',
-              }}
-              name="(home)/scan"
-            />
-            <Stack.Screen
-              name="notifications/openIdPresentation"
-              options={{
-                gestureEnabled: false,
-              }}
-            />
-            <Stack.Screen
-              name="notifications/openIdCredential"
-              options={{
-                gestureEnabled: false,
-              }}
-            />
-            <Stack.Screen
-              name="notifications/offlinePresentation"
-              options={{
-                gestureEnabled: false,
-              }}
-            />
-            <Stack.Screen name="credentials/index" options={headerNormalOptions} />
-            <Stack.Screen name="credentials/[id]/index" options={headerNormalOptions} />
-            <Stack.Screen name="credentials/[id]/attributes" options={headerNormalOptions} />
-            <Stack.Screen name="credentials/requestedAttributes" options={headerNormalOptions} />
-            <Stack.Screen name="menu/index" options={headerNormalOptions} />
-            <Stack.Screen name="menu/feedback" options={headerNormalOptions} />
-            <Stack.Screen name="menu/settings" options={headerNormalOptions} />
-            <Stack.Screen name="menu/about" options={headerNormalOptions} />
-            <Stack.Screen name="activity/index" options={headerNormalOptions} />
-            <Stack.Screen name="activity/[id]" options={headerNormalOptions} />
-            <Stack.Screen name="pinConfirmation" options={headerNormalOptions} />
-            <Stack.Screen name="pinLocked" options={headerNormalOptions} />
-            <Stack.Screen name="issuer" options={headerNormalOptions} />
-            <Stack.Screen name="pidSetup" />
-          </Stack>
-        </WithBackgroundPidRefresh>
+        <Stack screenOptions={{ headerShown: false }}>
+          <Stack.Screen
+            options={{
+              presentation: 'modal',
+            }}
+            name="(home)/scan"
+          />
+          <Stack.Screen
+            name="notifications/openIdPresentation"
+            options={{
+              gestureEnabled: false,
+            }}
+          />
+          <Stack.Screen
+            name="notifications/openIdCredential"
+            options={{
+              gestureEnabled: false,
+            }}
+          />
+          <Stack.Screen
+            name="notifications/offlinePresentation"
+            options={{
+              gestureEnabled: false,
+            }}
+          />
+          <Stack.Screen name="credentials/index" options={headerNormalOptions} />
+          <Stack.Screen name="credentials/[id]/index" options={headerNormalOptions} />
+          <Stack.Screen name="credentials/[id]/attributes" options={headerNormalOptions} />
+          <Stack.Screen name="credentials/requestedAttributes" options={headerNormalOptions} />
+          <Stack.Screen name="menu/index" options={headerNormalOptions} />
+          <Stack.Screen name="menu/feedback" options={headerNormalOptions} />
+          <Stack.Screen name="menu/settings" options={headerNormalOptions} />
+          <Stack.Screen name="menu/about" options={headerNormalOptions} />
+          <Stack.Screen name="activity/index" options={headerNormalOptions} />
+          <Stack.Screen name="activity/[id]" options={headerNormalOptions} />
+          <Stack.Screen name="pinConfirmation" options={headerNormalOptions} />
+          <Stack.Screen name="pinLocked" options={headerNormalOptions} />
+          <Stack.Screen name="issuer" options={headerNormalOptions} />
+          <Stack.Screen name="pidSetup" />
+        </Stack>
       </WalletJsonStoreProvider>
     </AgentProvider>
   )
