+++ packages/agent/src/invitation/index.ts
@@ -35,8 +35,8 @@ export {
   acquirePreAuthorizedAccessToken,
   resolveOpenId4VciOffer,
   getCredentialsForProofRequest,
-  shareProof,
   withTrustedCertificate,
   acquireAuthorizationCodeUsingPresentation,
 } from './handler'
+export { shareProof } from './shareProof'
 export * from './error'
